refactor(web): split backend into dispatch + bookkeeping + transport (Task 60a)
Non-destructive Option-A seam ahead of the generator. WebCommonGodotBackend.kt
now carries only the mechanical opcode routing, execution-mode guards, and codec
calls; Web-only stateful bookkeeping (property snapshots + read-your-write,
browser handle-kind tracking, free-time cache clearing) moves to
WebBackendBookkeeping.kt behind opcode-neutral hooks, and the js() bridge
externs move to WebBackendTransport.kt. Behavior-preserving: same snapshots,
handle kinds, guards, and flush points as before.

Verified: :web-runtime:compileKotlinWasmJs, release
(compileProductionExecutableKotlinWasmJs), generateWebGameplayCoverage
(blocking=0), and ktfmt all green.
